// src/schema/types.generated.ts
export type Account = {
  __typename?: 'Account';
  amount: Scalars['Float']['output'];
  barcode?: Maybe<Scalars['String']['output']>;
  callNumber?: Maybe<Scalars['String']['output']>;
  dateCreated?: Maybe<Scalars['DateTime']['output']>;
  dateUpdated?: Maybe<Scalars['DateTime']['output']>;
  dueDate?: Maybe<Scalars['DateTime']['output']>;
  feeFine?: Maybe<FeeFine>;
  feeFineId: Scalars['UUID']['output'];
  feeFineOwner?: Maybe<Scalars['String']['output']>;
  feeFineType?: Maybe<Scalars['String']['output']>;
  holdingsRecordId?: Maybe<Scalars['UUID']['output']>;
  id?: Maybe<Scalars['UUID']['output']>;
  instanceId?: Maybe<Scalars['UUID']['output']>;
  item?: Maybe<Item>;
  itemId?: Maybe<Scalars['UUID']['output']>;
  itemStatus?: Maybe<Status>;
  loan?: Maybe<CirculationLoan>;
  loanId?: Maybe<Scalars['UUID']['output']>;
  loanPolicyId?: Maybe<Scalars['UUID']['output']>;
  location?: Maybe<Scalars['String']['output']>;
  lostItemFeePolicyId?: Maybe<Scalars['UUID']['output']>;
  materialType?: Maybe<Scalars['String']['output']>;
  materialTypeId?: Maybe<Scalars['UUID']['output']>;
  overdueFinePolicyId?: Maybe<Scalars['UUID']['output']>;
  ownerId: Scalars['UUID']['output'];
  paymentStatus: Status;
  processId?: Maybe<Scalars['UUID']['output']>;
  remaining: Scalars['Float']['output'];
  returnedDate?: Maybe<Scalars['DateTime']['output']>;
  status?: Maybe<Status>;
  title?: Maybe<Scalars['String']['output']>;
  user?: Maybe<User>;
  userId: Scalars['UUID']['output'];
};

export type Address = {
  __typename?: 'Address';
  addressLine1?: Maybe<Scalars['String']['output']>;
  addressLine2?: Maybe<Scalars['String']['output']>;
  addressTypeId: Scalars['UUID']['output'];
  city?: Maybe<Scalars['String']['output']>;
  countryId?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  postalCode?: Maybe<Scalars['String']['output']>;
  primaryAddress?: Maybe<Scalars['Boolean']['output']>;
  region?: Maybe<Scalars['String']['output']>;
};

export type AlternativeTitle = {
  __typename?: 'AlternativeTitle';
  alternativeTitle?: Maybe<Scalars['String']['output']>;
  alternativeTitleTypeId?: Maybe<Scalars['UUID']['output']>;
};

export type Block = {
  __typename?: 'Block';
  blockBorrowing?: Maybe<Scalars['Boolean']['output']>;
  blockRenewals?: Maybe<Scalars['Boolean']['output']>;
  blockRequests?: Maybe<Scalars['Boolean']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  patronBlockConditionId?: Maybe<Scalars['UUID']['output']>;
};

export type BlockCondition = {
  __typename?: 'BlockCondition';
  blockBorrowing?: Maybe<Scalars['Boolean']['output']>;
  blockRenewals?: Maybe<Scalars['Boolean']['output']>;
  blockRequests?: Maybe<Scalars['Boolean']['output']>;
  id?: Maybe<Scalars['UUID']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  valueType?: Maybe<Scalars['String']['output']>;
};

export type BlockLimit = {
  __typename?: 'BlockLimit';
  condition?: Maybe<BlockCondition>;
  conditionId?: Maybe<Scalars['UUID']['output']>;
  id?: Maybe<Scalars['UUID']['output']>;
  patronGroupId?: Maybe<Scalars['UUID']['output']>;
  value?: Maybe<Scalars['Float']['output']>;
};
// ...
export type ChargeAmount = {
  __typename?: 'ChargeAmount';
  amount: Scalars['Float']['output'];
  isoCurrencyCode: Scalars['String']['output'];
};
// ...
export type AccountResolvers<ContextType = any, ParentType extends ResolversParentTypes['Account'] = ResolversParentTypes['Account']> = {
  amount?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
  barcode?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  callNumber?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  dateCreated?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  dateUpdated?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  dueDate?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  feeFine?: Resolver<Maybe<ResolversTypes['FeeFine']>, ParentType, ContextType>;
  feeFineId?: Resolver<ResolversTypes['UUID'], ParentType, ContextType>;
  feeFineOwner?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  feeFineType?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  holdingsRecordId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  id?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  instanceId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  item?: Resolver<Maybe<ResolversTypes['Item']>, ParentType, ContextType>;
  itemId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  itemStatus?: Resolver<Maybe<ResolversTypes['Status']>, ParentType, ContextType>;
  loan?: Resolver<Maybe<ResolversTypes['CirculationLoan']>, ParentType, ContextType>;
  loanId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  loanPolicyId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  location?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  lostItemFeePolicyId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  materialType?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  materialTypeId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  overdueFinePolicyId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  ownerId?: Resolver<ResolversTypes['UUID'], ParentType, ContextType>;
  paymentStatus?: Resolver<ResolversTypes['Status'], ParentType, ContextType>;
  processId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  remaining?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
  returnedDate?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  status?: Resolver<Maybe<ResolversTypes['Status']>, ParentType, ContextType>;
  title?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  user?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
  userId?: Resolver<ResolversTypes['UUID'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type AddressResolvers<ContextType = any, ParentType extends ResolversParentTypes['Address'] = ResolversParentTypes['Address']> = {
  addressLine1?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  addressLine2?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  addressTypeId?: Resolver<ResolversTypes['UUID'], ParentType, ContextType>;
  city?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  countryId?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  postalCode?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  primaryAddress?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  region?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type AlternativeTitleResolvers<ContextType = any, ParentType extends ResolversParentTypes['AlternativeTitle'] = ResolversParentTypes['AlternativeTitle']> = {
  alternativeTitle?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  alternativeTitleTypeId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BlockResolvers<ContextType = any, ParentType extends ResolversParentTypes['Block'] = ResolversParentTypes['Block']> = {
  blockBorrowing?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  blockRenewals?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  blockRequests?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  patronBlockConditionId?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BlockConditionResolvers<ContextType = any, ParentType extends ResolversParentTypes['BlockCondition'] = ResolversParentTypes['BlockCondition']> = {
  blockBorrowing?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  blockRenewals?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  blockRequests?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  id?: Resolver<Maybe<ResolversTypes['UUID']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  valueType?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type Resolvers<ContextType = any> = {
  Account?: AccountResolvers<ContextType>;
  Address?: AddressResolvers<ContextType>;
  AlternativeTitle?: AlternativeTitleResolvers<ContextType>;
  Block?: BlockResolvers<ContextType>;
  BlockCondition?: BlockConditionResolvers<ContextType>;
  BlockLimit?: BlockLimitResolvers<ContextType>;
  Campus?: CampusResolvers<ContextType>;
  Charge?: ChargeResolvers<ContextType>;
  ChargeAmount?: ChargeAmountResolvers<ContextType>;
  CheckIn?: CheckInResolvers<ContextType>;
  CirculationLoan?: CirculationLoanResolvers<ContextType>;
  CirculationLoanFeesAndFines?: CirculationLoanFeesAndFinesResolvers<ContextType>;
  CirculationNote?: CirculationNoteResolvers<ContextType>;
  Classification?: ClassificationResolvers<ContextType>;
  ClassificationType?: ClassificationTypeResolvers<ContextType>;
  Contributor?: ContributorResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  EffectiveCallNumberComponents?: EffectiveCallNumberComponentsResolvers<ContextType>;
  ElectronicAccess?: ElectronicAccessResolvers<ContextType>;
  EmailAddress?: GraphQLScalarType;
  FeeFine?: FeeFineResolvers<ContextType>;
  FixedDueDateSchedule?: FixedDueDateScheduleResolvers<ContextType>;
  Hold?: HoldResolvers<ContextType>;
  HoldingsNote?: HoldingsNoteResolvers<ContextType>;
  HoldingsRecord?: HoldingsRecordResolvers<ContextType>;
  HoldingsStatement?: HoldingsStatementResolvers<ContextType>;
  Identifier?: IdentifierResolvers<ContextType>;
  Instance?: InstanceResolvers<ContextType>;
  InstanceNote?: InstanceNoteResolvers<ContextType>;
  Institution?: InstitutionResolvers<ContextType>;
  Item?: ItemResolvers<ContextType>;
  ItemNote?: ItemNoteResolvers<ContextType>;
  Library?: LibraryResolvers<ContextType>;
  Loan?: LoanResolvers<ContextType>;
  LoanPolicy?: LoanPolicyResolvers<ContextType>;
  LoansPolicy?: LoansPolicyResolvers<ContextType>;
  Location?: LocationResolvers<ContextType>;
  ManualBlock?: ManualBlockResolvers<ContextType>;
  Metadata?: MetadataResolvers<ContextType>;
  Patron?: PatronResolvers<ContextType>;
  PatronGroup?: PatronGroupResolvers<ContextType>;
  Period?: PeriodResolvers<ContextType>;
  Personal?: PersonalResolvers<ContextType>;
  ProxyFor?: ProxyForResolvers<ContextType>;
  Publication?: PublicationResolvers<ContextType>;
  PublicationPeriod?: PublicationPeriodResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  ReceivingHistory?: ReceivingHistoryResolvers<ContextType>;
  ReceivingHistoryEntry?: ReceivingHistoryEntryResolvers<ContextType>;
  RenewalsPolicy?: RenewalsPolicyResolvers<ContextType>;
  RequestItem?: RequestItemResolvers<ContextType>;
  RequestPolicy?: RequestPolicyResolvers<ContextType>;
  Schedule?: ScheduleResolvers<ContextType>;
  ServicePoint?: ServicePointResolvers<ContextType>;
  Source?: SourceResolvers<ContextType>;
  Status?: StatusResolvers<ContextType>;
  Tag?: TagResolvers<ContextType>;
  Tags?: TagsResolvers<ContextType>;
  UUID?: GraphQLScalarType;
  User?: UserResolvers<ContextType>;
};